Separated gamepad color into its own files
